moves UI components to subdirectory
