+++ components/Header.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import React from "react";
-import Padding from "./Padding";
+import Padding from "./styled/Padding";
 
 const Header = () => {
   return (
