+++ components/Intro.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ShowType } from "../pages/shows/[show]";
-import Padding from "./Padding";
+import Padding from "./styled/Padding";
 import Shows from "./Shows";
 
 type ShowsType = ShowType[];
